scripts/requirements.txt
python-frontmatter>=1.1.0
jsonschema>=4.0.0
